+++ api/go.mod
@@ -1,12 +1,13 @@
-module undertaker
+module github.com/thinkocapo/undertaker/api
 
-go 1.12
+go 1.13
 
 require (
+	cloud.google.com/go/pubsub v1.3.1 // indirect
+	cloud.google.com/go/storage v1.10.0 // indirect
 	github.com/buger/goreplay v1.0.0
-	github.com/buger/jsonparser v1.0.0
 	github.com/google/uuid v1.1.1
 	github.com/joho/godotenv v1.3.0
-	github.com/mattn/go-sqlite3 v2.0.3+incompatible
 	github.com/shopspring/decimal v1.2.0
+	golang.org/x/exp v0.0.0-20200224162631-6cc2880d07d6 // indirect
 )
